+++ index.js
@@ -93,7 +93,7 @@ module.exports = (options, app) => {
 
   // If baseUrl and base are both set and home is not then lets set a better default
   if (app.options.base && options.baseUrl && !options.home) {
-    options.home = `/${path.relative(app.options.base, new URL(options.baseUrl).pathname)}`;
+    options.home = options.baseUrl;
   }
 
   // Add some custom containers
@@ -138,7 +138,10 @@ module.exports = (options, app) => {
     name: '@lando/vuepress-theme-default-plus',
     extends: '@vuepress/theme-default',
     alias: {
-      ...{'@theme/PageMeta.vue': path.resolve(__dirname, 'components', 'CustomPageMeta.vue')},
+      ...{
+        '@theme/NavbarBrand.vue': path.resolve(__dirname, 'components', 'CustomNavbarBrand.vue'),
+        '@theme/PageMeta.vue': path.resolve(__dirname, 'components', 'CustomPageMeta.vue'),
+      },
       ...options.alias,
     },
     define: {